Address Notification Opened and Actions
There are:
- OpenedEventActionType
- OpenedEvent
- OpenedEventAction

* Add missing actionId to OpenedEventAction

* Rename NotificationOpened.js to TypeScript file

- OSNotificationOpenedResult is same as OpenedEvent
- OSNotificationAction is same as OpenedEventAction
